removed skills + keywords from userWrapper
